finish matrix

+++ tools/inc/string_tools.h
@@ -1,5 +1,5 @@
 #include<vector>
 #include<string>
 namespace claris{
-std::vector<double> string2number(const std::string &line);
+    std::vector<long double> string2number(const std::string &line);
 }
