+++ tools/src/string_tools.cpp
@@ -2,8 +2,8 @@
 
 namespace claris{
 
-    std::vector<double> string2number(const std::string& str){
-        std::vector<double> res;
+    std::vector<long double> string2number(const std::string& str){
+        std::vector<long double> res;
         int begin = 0;
         for(int i=0;i<str.size();++i){
             if(str[i]==' '){
@@ -12,7 +12,7 @@ namespace claris{
                 begin = i;
                 ++i;
                 while(str[i]!=' '&&i<str.size())++i;
-                res.push_back(std::stod(str.substr(begin,i-begin)));
+                res.push_back(std::stold(str.substr(begin,i-begin)));
             }
         }
         return res;
